+++ packages/eslint-plugin-formatjs/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@4.4.0...eslint-plugin-formatjs@4.5.0) (2023-01-27)
+
+### Features
+
+* **eslint-plugin-formatjs:** capture literals in the logical expression ([#3975](https://github.com/formatjs/formatjs/issues/3975)) ([1cc638c](https://github.com/formatjs/formatjs/commit/1cc638c0a56cc19949327e55580e3a3e82a4116a))
+
+# [4.4.0](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@4.3.9...eslint-plugin-formatjs@4.4.0) (2023-01-26)
+
+### Features
+
+* **eslint-plugin-formatjs:** capture literals in the conditional expression ([#3974](https://github.com/formatjs/formatjs/issues/3974)) ([1deed4d](https://github.com/formatjs/formatjs/commit/1deed4d6978d1a48f4cf789920e8af8db29e7758))
+* **eslint-plugin-formatjs:** introduce no-useless-message linter rule ([#3973](https://github.com/formatjs/formatjs/issues/3973)) ([9ec4491](https://github.com/formatjs/formatjs/commit/9ec44911bc4b71616f4d5db3a7d7dbf4144a5a6a))
+
 ## [4.3.9](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@4.3.8...eslint-plugin-formatjs@4.3.9) (2022-12-27)
 
 ### Bug Fixes