+++ packages/swc-plugin-experimental/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/formatjs/formatjs/compare/@formatjs/swc-plugin-experimental@0.3.2...@formatjs/swc-plugin-experimental@0.3.3) (2023-01-26)
+
+**Note:** Version bump only for package @formatjs/swc-plugin-experimental
+
 ## [0.3.2](https://github.com/formatjs/formatjs/compare/@formatjs/swc-plugin-experimental@0.3.0...@formatjs/swc-plugin-experimental@0.3.2) (2022-12-27)
 
 **Note:** Version bump only for package @formatjs/swc-plugin-experimental